requirements.txt
streamlit==1.38.0
requests==2.32.3
python-dotenv==1.0.1
wordcloud==1.9.3
matplotlib==3.9.2
seaborn==0.13.2
pandas==2.2.2
pytest==8.3.3
pytest-mock==3.14.0
tenacity==9.0.0
